Modify messages/proposals/ban mechanism
Refactor a little bit the code when sending a proposal, receiving
an offer and answering. Check if we already know the peer. If
so, don't accept connection.

Accepting proposal condition:

the peer should be farther than 1 (not a direct connection). If we are not
currently connected to the peer (refer to the currData.connectedPeerIds).
If we don't have the ideal number of peers connections, we can accept it.
Otherwise, we choose randomly to accept or not 50/50.